remove untested gpu operators (#7172)
* skip sparse dot gpu tset. add sparse_nd_zeros gpu test

* remove sparse_retain gpu
